requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
autopep8==2.3.1
billiard==4.2.1
blinker==1.9.0
brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configargparse==1.7
cryptography==44.0.0
defusedxml==0.7.1
django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-redis==5.4.0
django-silk==5.3.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
docopt==0.6.2
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
flask==3.1.0
flask-cors==5.0.0
flask-login==0.6.3
flower==2.0.1
gevent==24.11.1
geventhttpclient==2.3.3
gprof2dot==2024.6.6
greenlet==3.1.1
gunicorn==23.0.0
humanize==4.11.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
locust==2.32.4
markdown==3.7
markupsafe==3.0.2
model-bakery==1.20.0
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
prometheus-client==0.21.1
prompt-toolkit==3.0.48
psutil==6.1.0
psycopg==3.2.3
pycodestyle==2.12.1
pycparser==2.22
pyjwt==2.10.1
pytest==8.3.4
pytest-django==4.9.0
pytest-watch==4.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.2.0
redis==5.2.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
ruff==0.8.2
setuptools==75.6.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
tornado==6.4.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
werkzeug==3.1.3
whitenoise==6.8.2
zope-event==5.0
zope-interface==7.2
